mlpipeline_utils/requirements.txt
apache-beam[gcp]==2.16.0
google-cloud-bigquery==1.17.1
six==1.12.0
tensorflow==1.14.0
tensorflow-data-validation==0.14.1
tensorflow-transform==0.14.0
tensorflow-model-analysis==0.14.0
kfp==0.1.32